Implemented restart logic for jobs that specify a timeout value.
By default, jobs never time out. But jobs can specify a timeout time span.
This allows work items that are "stuck" to be completed with a timeout error
message and restarted.

Note that the system has no way of knowing if the job *actually completed*,
so developers must be careful to specify a timeout that gives them reasonable
confidence that the job must really be done.